order: хоткеи, очередь, сессия, окна графиков и диаграмма производительности
Вкладка Order переведена на общие run-бар и таб-бар: Ctrl+R и Ctrl+Shift+R,
очередь order_queue, счётчик в статус-баре. Сессия и окна графиков сохраняются
в _last_order.json и _last_order_windows.json, результат выгружается в CSV по
правому клику.

Окна графиков устроены как у параметрики: окно знает свой вид и список
вкладок, чьи кривые в нём лежат, поэтому схемы сравниваются на одном графике.
Настройки отображения переехали с конфига на окно — масштаб оси в окне общий
для всех кривых, и хранить его на конфиге значило бы иметь столько мнений о
масштабе, сколько в окне вкладок.

Новый вид расчёта — performance: perfIntegrateKernel интегрирует ОДНУ
траекторию шагом узла, cudaEvent стоят вплотную вокруг cuLaunchKernel, так что
в замер не входят ни копирования, ни компиляция, ни прогревочные запуски.
Ошибки для оси X приходят из отдельного, не засекаемого прохода order-ядра по
той же сетке: на графике «время vs ошибка» это ровно та ошибка, ради которой
заплачено измеренное время.

В _config.csv замера пишется положение тумблера раздельной компиляции: без
встраивания шага в ядро та же задача считается в разы дольше (на Лоренце с
явным Эйлером — в 3.6 раза), и файлы, снятые при разном его положении, между
собой не сравнимы. Про дискретные отображения вкладка честно говорит, что
уточнять там нечего: шага у них нет.

Сюда же попали более ранние правки точности координат на графиках
(gpu_line_series / plot_view_2d / heatmap_view / analysis_session).

+++ U-CUDA/analysis_session.cpp
@@ -602,7 +602,7 @@ bool PhaseAnalysisSession::poll() {
                 per_var_out.assign(per_var_in.size(), {});
             for (size_t v = 0; v < per_var_in.size(); ++v) {
                 const FeaturePoints& fp = per_var_in[v];
-                std::vector<float>& buf = per_var_out[v];
+                std::vector<double>& buf = per_var_out[v];
                 double cumt = base;
                 const size_t n = fp.peaks.size() < fp.intervals.size()
                                ? fp.peaks.size() : fp.intervals.size();
@@ -610,8 +610,8 @@ bool PhaseAnalysisSession::poll() {
                 for (size_t p = 0; p < n; ++p) {
                     if (!std::isfinite(fp.peaks[p]) || !std::isfinite(fp.intervals[p])) continue;
                     cumt += fp.intervals[p];
-                    buf.push_back((float)cumt);
-                    buf.push_back((float)fp.peaks[p]);
+                    buf.push_back((double)cumt);
+                    buf.push_back((double)fp.peaks[p]);
                 }
                 const size_t cap_pairs = (size_t)std::max(1, continuation_peaks_cap);
                 const size_t cur_pairs = buf.size() / 2;
